organised into folders, and finished auto auth. Now its fully functional but still room for improvement
loginPage.dart will access 2 files, 1 is components.dart for its buttons and texts, and the other is checkin.dart for verifying the inputs with the database. Improvements exists on moving the materials from components.dart to a file closer so that it can be more easily modified. Another area is the auto auth, at the moment it has a little bit of lag and some initialization going around in the loginPage.dart Soon it will all be replaced by firebase as this php auth was all an experiment.